Functonality for deleting a RiSc through the backend (#418)
Adds endpoints for deletion of a RiSc. When the deletion endpoint is called for a specific RiSc, one of two things occur.
- If the RiSc has never been published, the draft branch of the RiSc is deleted and the RiSc disappears.
- If the RiSc has been published, the RiSc is deleted on its draft branch (branch is created if it does not exist)

Whenever a RiSc cannot be found on its draft branch, but the branch exists, it is considered marked for deletion. In these cases, the backend will retrieve the last published content and send this along with an updated status message `DeletionDraft`. If there also exists a PR from the draft branch to the default branch of the repository, then the status is instead `DeletionSentForApproval`.

The endpoint for creating PRs produces a different title and description based on if the RiSc has been deleted.